optimize the interrupt handler

+++ SRCDIR/FPGA/fpga.c
@@ -71,7 +71,7 @@ typedef struct isr_info_t {
 
 static isr_info_t g_isr_info[MAX_IRQ_NUM];
 
-int RegisterIsr(int bit_pos, ISR_HANDLER isr) {
+int RegisterIsr(UINT8 bit_pos, ISR_HANDLER isr) {
     if(bit_pos >= MAX_IRQ_NUM) {
         return 1;
     }
